\docType{package}
\name{genesearchr}
\alias{GeneSearchr}
\alias{genesearch}
\alias{genesearchr}
\alias{genesearchr-package}
\alias{package-genesearchr}
\title{Full-text search in NCBI RefSeq and/or UniProt-KB}
\description{
  This R package can search NCBI RefSeq and UniProt-KB from
  R using GGRNA and G-links Web API
}
\references{
  Yuki Naito & Hidemasa Bono (2012) GGRNA: an ultrafast,
  transcript-oriented search engine for genes and
  transcripts. Nucleic Acids Res. 40, W592-W596.

  Kazuharu Arakawa. G-Links is a rapid data "broker"
  service that collects and adds related information to a
  given gene (or gene set). http://link.g-language.org/
}

